Move Views and ViewModels to its own files
