+++ HackrNews/Stories Presentation/NewStoriesErrorView.swift
@@ -1,0 +1,9 @@
+//
+//  Copyright © 2020 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+
+public protocol NewStoriesErrorView {
+    func display(_ viewModel: NewStoriesErrorViewModel)
+}
