+++ HackrNews/Stories Presentation/NewStoriesLoadingViewModel.swift
@@ -1,0 +1,9 @@
+//
+//  Copyright © 2020 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+
+public struct NewStoriesLoadingViewModel {
+    public let isLoading: Bool
+}
